ApiUsuarios/requirements.txt
pandas
numpy
pydantic
google-cloud-storage
google-cloud-datastore
google-cloud-bigquery
pandas-gbq
scikit-surprise
scikit-learn
fastapi
uvicorn
gunicorn